frontend/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
PyMuPDF
google-generativeai==0.8.3
google-cloud-firestore
